+++ src/Domain/Event/EventStream.php
@@ -10,8 +10,6 @@
 namespace Slick\CQRSTools\Domain\Event;
 
 use ArrayIterator;
-use Countable;
-use IteratorAggregate;
 use Slick\CQRSTools\Event;
 
 /**
@@ -19,12 +17,17 @@ use Slick\CQRSTools\Event;
  *
  * @package Slick\CQRSTools\Domain\Event
 */
-final class EventStream implements IteratorAggregate, Countable
+final class EventStream implements Stream
 {
     /**
      * @var Event[]
      */
     private $events = [];
+
+    /**
+     * @var ArrayIterator
+     */
+    private $iterator;
 
     /**
      * Creates an Event Stream
@@ -40,6 +43,19 @@ final class EventStream implements IteratorAggregate, Countable
     }
 
     /**
+     * iterator
+     *
+     * @return ArrayIterator
+     */
+    private function iterator(): ArrayIterator
+    {
+        if (null === $this->iterator) {
+            $this->iterator = new ArrayIterator($this->events);
+        }
+        return $this->iterator;
+    }
+
+    /**
      * Adds an event to the stream
      *
      * @param Event|StoredEvent $event
@@ -49,6 +65,7 @@ final class EventStream implements IteratorAggregate, Countable
      */
     public function add(Event $event): EventStream
     {
+        $this->iterator = null;
         $event = $event instanceof StoredEvent ? $event->event() : $event;
         array_push($this->events, $event);
         return $this;
@@ -62,16 +79,6 @@ final class EventStream implements IteratorAggregate, Countable
     public function asArray(): array
     {
         return $this->events;
-    }
-
-    /**
-     * Retrieve an external iterator
-     *
-     * @return ArrayIterator
-     */
-    public function getIterator()
-    {
-        return new ArrayIterator($this->events);
     }
 
     /**
@@ -112,5 +119,56 @@ final class EventStream implements IteratorAggregate, Countable
     public function last(): ?Event
     {
         return end($this->events) ?: null;
+    }
+
+    /**
+     * Return the current element
+     *
+     * @return mixed Can return any type.
+     */
+    public function current()
+    {
+        return $this->iterator()->current();
+    }
+
+    /**
+     * Move forward to next element
+     *
+     * @return void Any returned value is ignored.
+     */
+    public function next()
+    {
+        $this->iterator()->next();
+    }
+
+    /**
+     * Return the key of the current element
+     *
+     * @return mixed scalar on success, or null on failure.
+     */
+    public function key()
+    {
+        return $this->iterator()->key();
+    }
+
+    /**
+     * Checks if current position is valid
+     *
+     * @return boolean The return value will be casted to boolean and then evaluated.
+     * Returns true on success or false on failure.
+     */
+    public function valid()
+    {
+        return $this->iterator()->valid();
+    }
+
+    /**
+     * Rewind the Iterator to the first element
+     *
+     * @return void Any returned value is ignored.
+     */
+    public function rewind()
+    {
+        $this->iterator()->rewind();
     }
 }