+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Event stream interface to allow other stream implementations
+### Change
+- `EventStream` to implement `Domain\Event\Stream` interface
 
 ## [v0.3.0] 2018-08-24
 ### Added