+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [v0.5.0] 2019-05-08
 ### Added
 - `ProgressStateProvider` interface to enable projectionist progress state notifications
 - `ProgressStateListener` interface that specifies the projectionist state changes listener
@@ -39,7 +41,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Initial composer definitions.
 - PHPSpec bootstrap
 
-[Unreleased]: https://github.com/slickframework/cqrs-tools/compare/v0.4.0...HEAD
+[Unreleased]: https://github.com/slickframework/cqrs-tools/compare/v0.5.0...HEAD
+[v0.5.0]: https://github.com/slickframework/cqrs-tools/compare/v0.4.0...v0.5.0
 [v0.4.0]: https://github.com/slickframework/cqrs-tools/compare/v0.3.0...v0.4.0
 [v0.3.0]: https://github.com/slickframework/cqrs-tools/compare/v0.2.0...v0.3.0
 [v0.2.0]: https://github.com/slickframework/cqrs-tools/compare/v0.1.0...v0.2.0